# profiles/apparmor: add some optimisations and tests