--- a/profiles/apparmor/template.go
+++ b/profiles/apparmor/template.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package apparmor // import "github.com/docker/docker/profiles/apparmor"
+package apparmor
 
 // NOTE: This profile is replicated in containerd and libpod. If you make a
 //       change to this profile, please make follow-up PRs to those projects so